Разделение js кода на модули
