+++ js/main.js
@@ -1,79 +1,3 @@
-const PICTURE_COUNT = 25;
-const AVATAR_COUNT = 6;
-const LIKE_MIN_COUNT = 15;
-const LIKE_MAX_COUNT = 200;
-const COMMENT_COUNT = 20;
-const COMMENT_LINES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-const DESCRIPTIONS = [
-  'Если смогу, я сделаю это. Конец истории.',
-  'Смейтесь как только умеете, любите столько, сколько живете.',
-  'Помните: вы единственный человек, который может наполнить ваш мир солнечным светом.',
-  'Я полностью уверена, что я — диснеевская принцесса, которую еще не придумали.',
-  'Не позволяйте кому-то затушить ваши искры только потому, что их свет сияет в чьих-то глазах.',
-  'Делайте в вашей жизни то, что меньше заставляет вас смотреть в свой телефон.',
-  'Улыбка — единственный тренд в моде, который актуален всегда.',
-  'Никогда не ищите свое счастье там, где вы его однажды потеряли.',
-  'Жизнь похожа на фотокамеру: вам просто нужно смотреть на нее с улыбкой.',
-  'Моя жизнь меняется, потому что меняю ее я.',
-  'Ни о чем не беспокойтесь. Потому что все лучшие умы на работе.',
-];
-const NAMES = ['Артем','Ульяна','Анастасия','Герствуд','Джо','Кэрри'];
+import {getPictures} from './create-data.js';
 
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (array) =>
-  array[getRandomInteger(0, array.length - 1)];
-
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
-
-const generateCommentId = createIdGenerator();
-
-const createMessage = () =>
-  Array.from({length: getRandomInteger(1, 2)}, () =>
-    getRandomArrayElement(COMMENT_LINES)
-  ).join(' ');
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
-  message: createMessage(),
-  name: getRandomArrayElement(NAMES),
-});
-
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
-  comments: Array.from(
-    {length: getRandomInteger(0, COMMENT_COUNT)},
-    createComment
-  ),
-});
-
-const getPictures = () =>
-  Array.from({length: PICTURE_COUNT}, (_, pictureIndex) =>
-    createPicture(pictureIndex + 1)
-  );
-
-getPictures();
+console.log(getPictures());
